requirements.txt
Django==1.6.5
beautifulsoup4==4.3.2
djangorestframework==2.3.14
gnureadline==6.3.3
ipdb==0.8
ipython==2.1.0
requests==2.3.0
wsgiref==0.1.2
